+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 Features:
   - MaimaiRoutes 支持指定 `with_curves` 参数
+  - 允许 `maimai.minfo()` 方法可空提供 `PlayerIdentifier` 参数
 
 ## 1.3.2 (2025-07-10)